Move I/O-related classes under org.spine3.io package.
Also, have separate io.file sub-package for file system routines and classes.
