+++ core-java-sample/src/main/java/org/spine3/server/storage/filesystem/FsCommandStorage.java
@@ -27,8 +27,8 @@ import java.io.File;
 import java.io.IOException;
 
 import static com.google.common.base.Preconditions.checkNotNull;
+import static org.spine3.io.file.FileUtil.createIfDoesNotExist;
 import static org.spine3.server.storage.filesystem.FsUtil.writeMessage;
-import static org.spine3.util.IoUtil.createIfDoesNotExist;
 
 /**
  * A storage for aggregate root events and snapshots based on the file system.
