+++ core-java/src/main/java/org/spine3/io/file/FileNameEscaper.java
@@ -17,7 +17,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.spine3.util;
+package org.spine3.io.file;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.CharMatcher;
