+++ core-java/src/main/java/org/spine3/protobuf/TypeToClassMap.java
@@ -25,10 +25,10 @@ import com.google.common.collect.ImmutableSet;
 import com.google.protobuf.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.spine3.io.IoUtil;
 import org.spine3.protobuf.error.UnknownTypeInAnyException;
 import org.spine3.type.ClassName;
 import org.spine3.type.TypeName;
-import org.spine3.util.IoUtil;
 
 import java.io.IOException;
 import java.io.InputStream;
